api/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
asyncio==3.4.3
attrs==23.1.0
bcrypt==4.1.2
bidict==0.22.1
blinker==1.7.0
Bottleneck==1.3.7
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
fastapi==0.105.0
fastapi-jwt==0.2.0
fastapi-socketio==0.0.10
financedatabase==2.2.1
financetoolkit==1.7.1
Flask==3.0.2
fonttools==4.47.0
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.352
langchain-community==0.0.5
langchain-core==0.1.2
langchain-experimental==0.0.47
langchain-google-genai==0.0.5
langsmith==0.0.72
llvmlite==0.41.1
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.8.2
multidict==6.0.4
mypy-extensions==1.0.0
numba==0.58.1
numexpr==2.8.8
numpy==1.26.2
openai==1.6.0
packaging==23.2
pandas==2.1.4
passlib==1.7.4
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==1.7.1
pymongo==4.6.1
pyngrok==7.0.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.1
python-jose==3.3.0
python-socketio==5.11.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
Redis-Sentinel-Url==1.0.1
requests==2.31.0
rq==1.15.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
xarray==2023.12.0
yarl==1.9.4
